+++ config/env.js
@@ -1,0 +1,4 @@
+/*eslint-env node*/
+module.exports = {
+  endpoint: 'http://localhost:3000'
+}
